requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.0

# Redis
redis==5.0.1

# OCR and Image Processing
pytesseract==0.3.10
Pillow==10.1.0
opencv-python==4.8.1.78
numpy<2.0.0

# HTTP client
httpx==0.25.2
requests==2.31.0

# Google AI
google-generativeai==0.3.2

# Environment management
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1